add jsconfig and craco & refactor project structure & remove redux
