+++ .prettierrc
@@ -1,0 +1,7 @@
+{
+  "arrowParens": "always",
+  "bracketSpacing": true,
+  "trailingComma": "all",
+  "singleQuote": true,
+  "jsxBracketSameLine": true
+}
