+++ craco.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  webpack: {
+    configure: {
+      target: 'electron-renderer',
+    },
+  },
+};
