+++ package.json
@@ -4,16 +4,14 @@
   "private": true,
   "main": "public/electron.js",
   "homepage": "./",
+  "cracoConfig": "craco.config.js",
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
+    "@craco/craco": "^5.7.0",
     "electron-is-dev": "^1.2.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-redux": "^7.2.0",
     "react-scripts": "3.4.1",
-    "redux": "^4.0.5"
+    "serialport": "^9.0.1"
   },
   "build": {
     "appId": "com.electron-react",
@@ -23,10 +21,11 @@
     "linux": {}
   },
   "scripts": {
-    "start": "concurrently \"react-scripts start\" \"wait-on http://localhost:3000 &&  electron .\"",
-    "build": "react-scripts build && electron-builder",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "start": "concurrently \"craco start\" \"wait-on http://localhost:3000 &&  electron .\"",
+    "build": "craco build && electron-builder",
+    "test": "craco test",
+    "eject": "craco eject",
+    "install": "electron-rebuild"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -47,6 +46,7 @@
     "concurrently": "^5.1.0",
     "electron": "^8.2.1",
     "electron-builder": "^22.4.1",
+    "electron-rebuild": "^2.0.3",
     "wait-on": "^4.0.2"
   }
 }
