+++ public/electron.js
@@ -1,21 +1,24 @@
-const { app, BrowserWindow } = require('electron')
-const path = require('path')
+const { app, BrowserWindow } = require('electron');
+const path = require('path');
 const isDev = require('electron-is-dev');
 
 function createWindow() {
-    // Create the browser window.     
-    win = new BrowserWindow({
-        width: 800,
-        height: 600,
-        webPreferences: {
-            preload: path.join(__dirname, './preload.js'),
-        }
-    })
+  // Create the browser window.
+  const win = new BrowserWindow({
+    width: 800,
+    height: 600,
+    webPreferences: {
+      preload: path.join(__dirname, './preload.js'),
+      nodeIntegration: true,
+    },
+  });
 
-    //load from dev server or from builded file
-    win.loadURL(isDev ?
-        'http://localhost:3000' :
-        `file://${path.join(__dirname, './index.html')}`
-    );
+  //load from dev server or from builded file
+  win.loadURL(
+    isDev
+      ? 'http://localhost:3000'
+      : `file://${path.join(__dirname, './index.html')}`,
+  );
 }
-app.on('ready', createWindow)
+
+app.on('ready', createWindow);
