+++ src/components/App/index.jsx
@@ -1,8 +1,17 @@
-import React from 'react';
-import logo from './logo.svg';
-import './App.css';
+import React, { useEffect } from 'react';
+import logo from 'assets/images/logo.svg';
+import './styles.css';
+
+import { getPorts } from 'utils/serialport';
 
 function App() {
+  useEffect(() => {
+    (async () => {
+      const ports = await getPorts();
+      console.log('PORTS ->', ports);
+    })();
+  }, []);
+
   return (
     <div className="App">
       <header className="App-header">
@@ -14,8 +23,7 @@ function App() {
           className="App-link"
           href="https://reactjs.org"
           target="_blank"
-          rel="noopener noreferrer"
-        >
+          rel="noopener noreferrer">
           Learn React
         </a>
       </header>
