--- src/reducers/reducer.js
@@ -1,8 +1,0 @@
-export default function reducer(state = initialState, action) {
-    //write your reducer here
-    return state;
-}
-
-const initialState = {
-    //write your initial state
-};
